fix: Properly handle status message on older kernels
The status message has evolved, with fields being added in subsequent
kernel versions. The parsing of this message excepts a recent version of
it, making it fail to parse on older kernels.

This is fixed by putting the newer fields behind an option. On parsing,
those are set only if the received status payload is big enough.