feat: parse AUDIT_REPLACE properly
The AUDIT_REPLACE message payload is a uint32 number, which is not an
ascii string. This causes parsing issues when such a message is parsed.

There is two solutions for this afaict:

- Parse the AUDIT_REPLACE message on its own, making minimal changes.
- Change the AuditMessage Event/Other enums to store a Vec<u8> instead
  of a String.

I picked the first one, to have minimal impact on existing code.
However, the second solution imho would be a much better solution.
This crate should not expect every message to a utf-8, which is a weird
assumption. Its much better to let the caller decide how to parse the
payloads, and not have errors when messages have non utf-8 compatible
bytes.